+++ packages/color/src/color.ts
@@ -1,0 +1,25 @@
+interface Color {
+    name: string;
+    color: string;
+}
+
+const colors: Color[] = [
+    {
+        name: "primary",
+        color: "#4FB058"
+    },
+    {
+        name: "secondary",
+        color: "#52525b"
+    },
+    {
+        name: "danger",
+        color: "#b91c1c"
+    },
+    {
+        name: "warning",
+        color: "#ca8a04"
+    }
+];
+
+export { colors };