+++ packages/theme/src/components/button.ts
@@ -4,11 +4,11 @@ const button = tv({
     base: "font-medium px-6 py-3 rounded-lg enabled:hover:opacity-80 transition-opacity duration-100 disabled:opacity-80",
     variants: {
         color: {
-            primary: "bg-[#4FB055]",
-            secondary: "bg-zinc-600",
-            outline: "border-[#4FB055] border-2 text-white",
-            danger: "bg-red-700",
-            warning: "bg-yellow-600",
+            primary: "bg-primary",
+            secondary: "bg-secondary",
+            outline: "border-primary border-2 text-white",
+            danger: "bg-danger",
+            warning: "bg-warning",
             ghost: "bg-transparent",
             link: "hover:underline"
         }