+++ packages/components/button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @kokoro-ui/button
 
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @kokoro-ui/theme@0.0.7
+
 ## 0.0.13
 
 ### Patch Changes